UIEXT-1346: Emit scroll update events for large rows
Increased the buffer so that the end index of the first scroll event is
at least 2. There exists a piece of code in the Recycle scroller which
prevents an update event to be emitted when the user scrolls less than a
row height. So when scrolling to the second row it should already be
loaded (i.e. the first emitted end index should be 2). The change can be
reverted when https://github.com/Akryum/vue-virtual-scroller/issues/834
is resolved.

UIEXT-1346 (Adjust max table height logic to account for row height)